json storage for timeline
