+++ core/Cargo.toml
@@ -27,6 +27,5 @@ toml = { workspace = true }
 tar = "0.4"
 flate2 = "1.0"
 chrono = { version = "0.4", features = ["serde"] }
-rusqlite = { version = "0.32", features = ["bundled"] }
 
 # NO Tauri dependencies - platform-agnostic!
