+++ core/src/lib.rs
@@ -3,6 +3,7 @@ pub mod error;
 pub mod ai;
 pub mod config;
 pub mod executor;
+pub mod timeline;
 pub mod tools;
 pub mod export;
