+++ server/src/main.rs
@@ -5,7 +5,7 @@ mod routes;
 
 use axum::{routing::get, Router};
 use reprod_core::{
-    executor::timeline::SqliteTimeline, Config, RExecutor, ToolExecutor, ToolRegistry,
+    executor::timeline::JsonTimeline, Config, RExecutor, ToolExecutor, ToolRegistry,
 };
 use std::path::PathBuf;
 use std::sync::Arc;
@@ -50,10 +50,10 @@ async fn main() {
     let tool_executor = Arc::new(ToolExecutor::new(tool_registry.clone()));
 
     // Initialize timeline storage
-    let timeline_db_path = std::env::temp_dir().join("reprod").join("timeline.db");
-    let timeline = SqliteTimeline::new(timeline_db_path).unwrap_or_else(|error| {
+    let timeline_file_path = std::env::temp_dir().join("reprod").join("timeline.ndjson");
+    let timeline = JsonTimeline::new(timeline_file_path).unwrap_or_else(|error| {
         tracing::warn!("Failed to initialize timeline storage: {}. Using in-memory timeline.", error);
-        SqliteTimeline::new_in_memory().expect("Failed to create in-memory timeline")
+        JsonTimeline::new_in_memory().expect("Failed to create in-memory timeline")
     });
     let timeline = Arc::new(timeline);
     tracing::info!("Timeline storage initialized");
